refactor(cli): embed the worker starters with a Bun macro
`workers new` baked its starter files into the binary through a
`SUPABASE_WORKER_STACKS` define, falling back to reading `stacks/` off disk
when running from source. Because a missed define only broke a released
binary, that shape needed a build-time define script plumbed into every
`bun build` site, a runtime completeness assertion, an error class declaring
its own actionability, a module-level cache, and a unit suite guarding it.

Expand `stacks/` through a Bun macro instead. The macro runs while
`worker-stacks.ts` is transpiled and its return value is inlined as a
literal, so a compiled binary carries the content with nothing to pass at a
build site and no directory to find at runtime. Bun expands macros in the
runtime transpiler too, so running from source behaves the same; Vitest does
not implement them and degrades to calling the function against the source
tree, which is why the path comes from `import.meta.url` rather than Bun's
`import.meta.dir`.

Discovery stays directory-driven — a new runtime is a new directory plus its
`WORKER_RUNTIMES` entry — and the completeness check moves inside the macro,
where it fails the build rather than the binary. Bun reports a throwing macro
as one it could not coerce to AST, so the reason is logged first to keep the
diagnostic legible.
