+++ apps/cli/scripts/build.ts
@@ -5,7 +5,6 @@ import path from "node:path";
 import process from "node:process";
 import { parseArgs } from "node:util";
 import { bundleServeMainTemplate } from "../src/shared/functions/serve-main-bundler.ts";
-import { workerStacksDefine } from "./worker-stacks-define.ts";
 import { darwinBinariesForShell, MACOS_IDENTIFIERS } from "./macos-signing.ts";
 
 const MUSL_TARGETS = [
@@ -98,11 +97,6 @@ const goSource = path.resolve(root, "apps/cli-go");
 const serveMainTemplateDefine = `--define=SUPABASE_FUNCTIONS_SERVE_MAIN_TEMPLATE=${JSON.stringify(
   await bundleServeMainTemplate(),
 )}`;
-// Baked in so `workers new` can scaffold from a binary with no `stacks/`
-// directory beside it. Computed once, before any target is compiled, so an
-// incomplete stacks directory fails the build rather than shipping a binary
-// that scaffolds nothing.
-const stacksDefine = workerStacksDefine();
 const posthogBuildDefines = [
   `--define=process.env.SUPABASE_CLI_POSTHOG_KEY=${JSON.stringify(process.env.POSTHOG_API_KEY ?? "")}`,
   `--define=process.env.SUPABASE_CLI_POSTHOG_HOST=${JSON.stringify(process.env.POSTHOG_ENDPOINT ?? "")}`,
@@ -157,7 +151,6 @@ async function buildTarget(target: (typeof TARGETS)[number]) {
     `--define=process.env.SUPABASE_CLI_VERSION=${JSON.stringify(version)}`,
     `--define=SUPABASE_LIBC=${JSON.stringify(libc)}`,
     serveMainTemplateDefine,
-    stacksDefine,
     ...posthogBuildDefines,
     `--outfile=${outfile}`,
   ]);
@@ -307,7 +300,6 @@ async function buildMuslBinaries() {
         `--define=process.env.SUPABASE_CLI_VERSION=${JSON.stringify(version)}`,
         `--define=SUPABASE_LIBC=${JSON.stringify(libc)}`,
         serveMainTemplateDefine,
-        stacksDefine,
         ...posthogBuildDefines,
         `--outfile=${outfile}`,
       ]);
